# openapi.yaml
openapi: "3.0.3"
info:
  title: Zoro API
  description: Research orchestration API with knowledge graph integration.
  version: "1.0"

servers:
  - url: http://localhost:8080

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Health check
      tags: [health]
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: object
                additionalProperties:
                  type: string

  /api/research:
    post:
      operationId: startResearch
      summary: Start research session
      description: Creates a new research session for the given query
      tags: [research]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/StartResearchRequest"
      responses:
        "200":
          description: Session ID
          content:
            application/json:
              schema:
                type: object
                properties:
                  id:
                    type: string
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/research/{id}:
    get:
      operationId: getResearchSession
      summary: Get research session
      description: Returns the research session with the given ID
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ResearchSession"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/research/{id}/stream:
    get:
      operationId: streamResearchEvents
      summary: Stream research events
      description: Returns an SSE stream of research progress events
      tags: [research]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Session ID
      responses:
        "200":
          description: OK
          content:
            text/event-stream:
              schema:
                $ref: "#/components/schemas/SSEEvent"
        "404":
          description: Session not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/knowledge/search:
    get:
      operationId: searchKnowledge
      summary: Search knowledge graph
      description: Searches for facts in the knowledge graph
      tags: [knowledge]
      parameters:
        - name: q
          in: query
          required: true
          schema:
            type: string
          description: Search query
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SearchFactsResponse"
        "400":
          description: Missing query
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: Search failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/knowledge/graph:
    get:
      operationId: getKnowledgeGraph
      summary: Get knowledge graph
      description: Returns nodes and edges from the knowledge graph
      tags: [knowledge]
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 300
          description: Max number of nodes
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/GraphData"
        "500":
          description: Fetch failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /api/knowledge/node/{id}:
    get:
      operationId: getNodeDetail
      summary: Get node detail
      description: Returns a node and its neighbors from the knowledge graph
      tags: [knowledge]
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
          description: Node ID
        - name: depth
          in: query
          schema:
            type: integer
            default: 1
          description: Neighbor depth
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NodeDetail"
        "500":
          description: Node not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    ErrorResponse:
      type: object
      additionalProperties:
        type: string

    StartResearchRequest:
      type: object
      required: [query]
      properties:
        query:
          type: string
          example: "What is quantum computing?"

    Entity:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        type:
          type: string
        summary:
          type: string

    Relation:
      type: object
      properties:
        uuid:
          type: string
        source_uuid:
          type: string
        target_uuid:
          type: string
        type:
          type: string
        fact:
          type: string

    SearchResult:
      type: object
      properties:
        title:
          type: string
        url:
          type: string
        snippet:
          type: string

    TimelineEvent:
      type: object
      properties:
        type:
          type: string
        message:
          type: string
        timestamp:
          type: string
        data:
          type: object

    SSEEvent:
      type: object
      properties:
        type:
          type: string
        data:
          type: object

    ResearchSession:
      type: object
      properties:
        id:
          type: string
        query:
          type: string
        status:
          type: string
        summary:
          type: string
        created_at:
          type: string
        results:
          type: array
          items:
            $ref: "#/components/schemas/SearchResult"
        entities:
          type: array
          items:
            $ref: "#/components/schemas/Entity"
        relations:
          type: array
          items:
            $ref: "#/components/schemas/Relation"
        timeline:
          type: array
          items:
            $ref: "#/components/schemas/TimelineEvent"

    Fact:
      type: object
      properties:
        uuid:
          type: string
        name:
          type: string
        fact:
          type: string
        source_node:
          $ref: "#/components/schemas/Entity"
        target_node:
          $ref: "#/components/schemas/Entity"

    SearchFactsResponse:
      type: object
      properties:
        facts:
          type: array
          items:
            $ref: "#/components/schemas/Fact"

    GraphNode:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        type:
          type: string
        summary:
          type: string

    GraphEdge:
      type: object
      properties:
        id:
          type: string
        source:
          type: string
        target:
          type: string
        type:
          type: string
        fact:
          type: string
        weight:
          type: number

    GraphData:
      type: object
      properties:
        nodes:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"

    NodeDetail:
      type: object
      properties:
        node:
          $ref: "#/components/schemas/GraphNode"
        neighbors:
          type: array
          items:
            $ref: "#/components/schemas/GraphNode"
        edges:
          type: array
          items:
            $ref: "#/components/schemas/GraphEdge"